sourcing/dags/requirements.txt
apache-beam==2.36.0
apache-beam[gcp]==2.36.0
apache-airflow-providers-apache-beam==3.1.0
apache-airflow-providers-snowflake
pymssql
pymysql
psycopg2
python-dotenv
paramiko
ssh2-python==1.0.0
google-cloud-storage
gcsfs==2022.8.2
apitools
google-apitools
python-crontab==2.5.1
google-cloud-compute==1.19.2